Changed Speed Value to Float
As stated in title

+++ Assets/Scripts/AttackBehaviour.cs
@@ -9,7 +9,7 @@ public abstract class AttackBehaviour : MonoBehaviour
     protected DateTime timer;
     protected abstract void Start();
     protected abstract void Update();
-    public abstract void SetSpeed(int speed);
+    public abstract void SetSpeed(float speed);
     public void SetDirection(Vector3 direction)
     {
         this.direction = direction;
